src/requirements.txt
numpy==1.24.2
pillow==9.5.0
opencv-python==4.7.0.72
ttkthemes==3.2.2